\name{constant}
\alias{constant}
\title{constant prediction}
\usage{
  constant(c)
}
\arguments{
  \item{c}{the prediction value}
}
\value{
  a predictor function, a function of a date vector
}
\description{
  constant predictions
}
\details{
  predict the same number always.
}
\author{
  Barry S Rowlingson
}

