\name{dailyPoisson}
\alias{dailyPoisson}
\title{Constant Poisson Rate generator}
\usage{
  dailyPoisson(rate)
}
\arguments{
  \item{rate}{mean number of cases per day}
}
\value{
  a poisson rate generator function
}
\description{
  Independent Identical Poisson rate generator
}
\author{
  Barry S Rowlingson
}

