% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extendData}
\alias{extendData}
\title{Generate extended covariate matrix}
\usage{
extendData(dataset, formula, npredict)
}
\arguments{
\item{dataset}{input data set}

\item{formula}{of the form count~date}

\item{npredict}{number of extra rows to add on}
}
\value{
a data frame of covariates for the model
}
\description{
Extend a data set for prediction
}
\details{
given a time-count dataset, return the covariate matrix with added observations after the last one
with NA as observed value
}
\author{
Barry S Rowlingson
}

