% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mape}
\alias{mape}
\title{mape}
\usage{
mape(forecast, actual)
}
\arguments{
\item{forecast}{forecasted value}

\item{actual}{actual value}
}
\value{
the MAPE
}
\description{
mean absolute percentage errors
}
\details{
compute the mape
}
\author{
Barry Rowlingson
}

