% Generated by roxygen2 (4.0.1): do not edit by hand
\name{part1}
\alias{part1}
\title{part1 - fitting}
\usage{
part1(dataset, formula, modelName = c("Static", "DynamicInterceptCosSin",
  "DynamicIntercept", "DynamicCosSin"))
}
\arguments{
\item{dataset}{}

\item{formula}{}

\item{modelName}{}
}
\value{
the model object
}
\description{
Fit model to historical data
}
\details{
Given some date-count data and a model name, return the fitted model parameters
}
\author{
Barry S Rowlingson
}

