% Generated by roxygen2 (4.0.1): do not edit by hand
\name{part2}
\alias{part2}
\title{part2 - prediction}
\usage{
part2(model, dataset, formula, npredict = 28, nsims = 1000, nmax = 60,
  quantiles = c(0.5, 0.95, 0.99))
}
\arguments{
\item{model}{returned from part1}

\item{dataset}{data frame with day and count columns}

\item{formula}{of the form count~date}

\item{npredict}{number of days ahead to predict}

\item{nsims}{number of simulations}
}
\value{
means and confidence intervals
}
\description{
make predictions from dynamic model
}
\details{
given a model from part1, and a dataset, extend that dataset and predict
for those new days by simulation
}
\author{
Barry S Rowlingson
}

