\name{stayCondition}
\alias{stayCondition}
\title{Return a stay-time function generator}
\usage{
  stayCondition(conditions)
}
\arguments{
  \item{conditions}{data frame of condition info}
}
\value{
  stay-time function generator
}
\description{
  Stay-time function generator for conditions
}
\author{
  Barry S Rowlingson
}

