\name{stayIIDExponential}
\alias{stayIIDExponential}
\title{Constant stay-time generator}
\usage{
  stayIIDExponential(meanStay)
}
\arguments{
  \item{meanStay}{mean stay time in days}
}
\value{
  a stay-time generator
}
\description{
  Constant stay-time generator
}
\author{
  Barry S Rowlingson
}

