% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dailyPoisson}
\alias{dailyPoisson}
\title{Constant Poisson Rate generator}
\usage{
dailyPoisson(rate)
}
\arguments{
\item{rate}{mean number of cases per day}
}
\value{
a poisson rate generator function
}
\description{
Independent Identical Poisson rate generator
}
\author{
Barry S Rowlingson
}

