% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hospital}
\alias{hospital}
\title{Make a hospital object}
\usage{
hospital(name, medicalWards, surgicalWards)
}
\arguments{
\item{name}{Hospital name}

\item{medicalWards}{Medical ward data frame - Name and Size}

\item{surgicalWards}{Surgical ward data frame - Name and Size}
}
\value{
A hospital object
}
\description{
Make a hospital object
}
\author{
Barry S Rowlingson
}

