\name{todBetween}
\alias{todBetween}
\title{uniform time-of-day between hours}
\usage{
  todBetween(hlo, hhi)
}
\arguments{
  \item{hlo}{early hour}

  \item{hhi}{late hour - must be greater than hlo}
}
\value{
  a time-of-day generator.
}
\description{
  uniform time-of-day between given hours
}
\author{
  Barry S Rowlingson
}

